+++ client/Assets/BeamSemiAutomatedTester/Scripts/Editor/UI/Components/ActionBarVisualElement/ActionBarVisualElement.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using UnityEngine.UIElements;
-using UnityEditor.UIElements;
 
 namespace Beamable.BSAT.Editor.UI.Components
 {
@@ -12,18 +11,23 @@ namespace Beamable.BSAT.Editor.UI.Components
 		public event Action OnDeleteTestSceneButtonPressed;
 		public event Action OnGenerateReportButtonPressed;
 		public event Action OnOpenMainMenuSceneButtonPressed;
+		public event Action OnSetupBuildSettingsButtonPressed;
 
 		private Button _scanButton;
 		private Button _createTestSceneButton;
 		private Button _deleteTestSceneButton;
 		private Button _generateReportButton;
 		private Button _openMainMenuSceneButton;
+		private Button _setupBuildSettingsButton;
 
 		public new class UxmlFactory : UxmlFactory<ActionBarVisualElement, UxmlTraits> { }
+
 		public new class UxmlTraits : VisualElement.UxmlTraits
 		{
 			UxmlStringAttributeDescription customText = new UxmlStringAttributeDescription
-			{ name = "custom-text", defaultValue = "nada" };
+			{
+				name = "custom-text", defaultValue = "nada"
+			};
 
 			public override IEnumerable<UxmlChildElementDescription> uxmlChildElementsDescription
 			{
@@ -46,11 +50,11 @@ namespace Beamable.BSAT.Editor.UI.Components
 			_scanButton = Root.Q<Button>("scan");
 			_scanButton.clickable.clicked -= HandleScanButtonPressed;
 			_scanButton.clickable.clicked += HandleScanButtonPressed;
-			
+
 			_createTestSceneButton = Root.Q<Button>("createTestScene");
 			_createTestSceneButton.clickable.clicked -= HandleCreateTestSceneButton;
 			_createTestSceneButton.clickable.clicked += HandleCreateTestSceneButton;
-			
+
 			_deleteTestSceneButton = Root.Q<Button>("deleteTestScene");
 			_deleteTestSceneButton.clickable.clicked -= HandleDeleteTestSceneButton;
 			_deleteTestSceneButton.clickable.clicked += HandleDeleteTestSceneButton;
@@ -58,10 +62,14 @@ namespace Beamable.BSAT.Editor.UI.Components
 			_generateReportButton = Root.Q<Button>("generateReport");
 			_generateReportButton.clickable.clicked -= HandleReportGenerateButton;
 			_generateReportButton.clickable.clicked += HandleReportGenerateButton;
-			
+
 			_openMainMenuSceneButton = Root.Q<Button>("openMainMenuScene");
 			_openMainMenuSceneButton.clickable.clicked -= HandleOpenMainMenuSceneButton;
 			_openMainMenuSceneButton.clickable.clicked += HandleOpenMainMenuSceneButton;
+			
+			_setupBuildSettingsButton = Root.Q<Button>("setupBuildSetting");
+			_setupBuildSettingsButton.clickable.clicked -= HandleSetupBuildSettingsButton;
+			_setupBuildSettingsButton.clickable.clicked += HandleSetupBuildSettingsButton;
 		}
 
 		private void HandleScanButtonPressed() => OnScanButtonPressed?.Invoke();
@@ -69,5 +77,6 @@ namespace Beamable.BSAT.Editor.UI.Components
 		private void HandleDeleteTestSceneButton() => OnDeleteTestSceneButtonPressed?.Invoke();
 		private void HandleReportGenerateButton() => OnGenerateReportButtonPressed?.Invoke();
 		private void HandleOpenMainMenuSceneButton() => OnOpenMainMenuSceneButtonPressed?.Invoke();
+		private void HandleSetupBuildSettingsButton() => OnSetupBuildSettingsButtonPressed?.Invoke();
 	}
 }