+++ client/Packages/com.beamable/Editor/UI/Buss/ThemeManager/StylePropertyVisualElement/StylePropertyModel.cs
@@ -55,8 +55,7 @@ namespace Beamable.Editor.UI.Components
 
 		public void GetResult(out IBussProperty bussProperty, out VariableDatabase.PropertyReference propertyReference)
 		{
-			VariablesDatabase.TryGetProperty(PropertyProvider,
-											 StyleRule, out IBussProperty property,
+			VariablesDatabase.TryGetProperty(PropertyProvider, StyleRule, out IBussProperty property,
 											 out VariableDatabase.PropertyReference variableSource);
 
 			bussProperty = property;
@@ -157,6 +156,23 @@ namespace Beamable.Editor.UI.Components
 			var value = options.FindIndex(option => option.Equals(variableName));
 			value = Mathf.Clamp(value, 0, options.Count - 1);
 			return value;
+		}
+
+		public void OnPropertyChanged(IBussProperty property)
+		{
+			if (!StyleRule.HasProperty(PropertyProvider.Key))
+			{
+				StyleRule.TryAddProperty(PropertyProvider.Key, property);
+			}
+			else
+			{
+				StyleRule.GetPropertyProvider(PropertyProvider.Key).SetProperty(property);
+			}
+
+#if UNITY_EDITOR
+			EditorUtility.SetDirty(StyleSheet);
+			AssetDatabase.SaveAssets();
+#endif
 		}
 	}
 }