+++ client/Packages/com.beamable/Editor/UI/Buss/ThemeManager/ThemeModel.cs
@@ -27,21 +27,11 @@ namespace Beamable.Editor.UI.Buss
 
 		public abstract BussElement SelectedElement { get; set; }
 
-		protected abstract List<BussStyleSheet> StyleSheets { get; }
+		protected abstract List<BussStyleSheet> SceneStyleSheets { get; }
 
 		public abstract Dictionary<BussStyleRule, BussStyleSheet> FilteredRules { get; }
 
-		public IEnumerable<BussStyleSheet> WritableStyleSheets
-		{
-			get
-			{
-#if BEAMABLE_DEVELOPER
-				return StyleSheets ?? Enumerable.Empty<BussStyleSheet>();
-#else
-				return StyleSheets?.Where(s => !s.IsReadOnly) ?? Enumerable.Empty<BussStyleSheet>();
-#endif
-			}
-		}
+		public abstract List<BussStyleSheet> WritableStyleSheets { get; }
 
 		public VariableDatabase VariablesDatabase => BussConfiguration.OptionalInstance.Value.VariableDatabase;
 		public PropertySourceDatabase PropertyDatabase { get; } = new PropertySourceDatabase();
@@ -54,11 +44,6 @@ namespace Beamable.Editor.UI.Buss
 		public void NavigationElementClicked(BussElement element)
 		{
 			Selection.activeGameObject = Selection.activeGameObject == element.gameObject ? null : element.gameObject;
-		}
-
-		public void OnFocus()
-		{
-			Change?.Invoke();
 		}
 
 		#region Action bar buttons' actions
@@ -115,7 +100,7 @@ namespace Beamable.Editor.UI.Buss
 
 		public void OnCopyButtonClicked()
 		{
-			List<BussStyleSheet> readonlyStyles = StyleSheets.Where(styleSheet => styleSheet.IsReadOnly).ToList();
+			List<BussStyleSheet> readonlyStyles = SceneStyleSheets.Where(styleSheet => styleSheet.IsReadOnly).ToList();
 			OpenCopyMenu(readonlyStyles);
 		}